+++ index.html
@@ -36,7 +36,7 @@
       <section class="img-upload">
         <div class="img-upload__wrapper">
           <h2 class="img-upload__title  visually-hidden">Загрузка фотографии</h2>
-          <form class="img-upload__form" id="upload-select-image" method="post" enctype="multipart/form-data"
+          <form class="img-upload__form" action=" https://22.javascript.pages.academy/kekstagram" id="upload-select-image" method="post" enctype="multipart/form-data"
             autocomplete="off">
 
             <!-- Изначальное состояние поля для загрузки изображения -->
@@ -136,7 +136,7 @@
                 </fieldset>
 
                 <!-- Кнопка для отправки данных на сервер -->
-                <button type="submit" class="img-upload__submit" id="upload-submit">Опубликовать</button>
+                <button type="submit"  class="img-upload__submit" id="upload-submit">Опубликовать</button>
               </div>
             </div>
           </form>
