+++ js/bigPicture.js
@@ -8,6 +8,8 @@ const POPUP_SOCIAL_LIKES = POPUP_SOCIAL_HEADER.querySelector('.social__likes .li
 const POPUP_COMMENS_COUNT = POPUP_SOCIAL.querySelector('.social__comment-count .comments-count');
 const POPUP_SOCIAL_COMMENTS = POPUP_SOCIAL.querySelector('.social__comments');
 
+
+
 const COMMENTS_LOADER = document.querySelector('.comments-loader ');
 const SOCIAL_COMMENT_COUNT = document.querySelector('.social__comment-count');
 const BODY = document.querySelector('body');
@@ -112,247 +114,6 @@ const getObjectDataPic = pictures => (evt) => {
   }
 
 }
-//6 level
-
-const UPLOAD_FILE = document.querySelector('#upload-file');
-const UPLOAD_FILE_CANCEL = document.querySelector('.img-upload__cancel');
-const UPLOAD_OVERLAY = document.querySelector('.img-upload__overlay');
-
-
-
-UPLOAD_FILE_CANCEL.addEventListener('click', function () {
-  UPLOAD_OVERLAY.classList.add('hidden');
-  BODY.classList.remove('modal-open');
-  // COMMENTS_LOADER.classList.remove('hidden');
-  // SOCIAL_COMMENT_COUNT.classList.remove('hidden');
-});
-
-UPLOAD_FILE_CANCEL.addEventListener('keydown', (evt) => {
-  if (evt.key === ('Escape' || 'Esc')) {
-    //evt.preventDefault();
-    UPLOAD_OVERLAY.classList.add('hidden');
-    BODY.classList.remove('modal-open');
-  }
-});
-UPLOAD_FILE.addEventListener('change', function () {
-  UPLOAD_OVERLAY.classList.remove('hidden');
-  BODY.classList.add('modal-open');
-});
-
-
-const IMG_UPLOAD_EFFECT_LEVEL = document.querySelector('.img-upload__effect-level');
-const SLIDER_ELEMENT = IMG_UPLOAD_EFFECT_LEVEL.querySelector('.effect-level__slider');
-const UP_SCALE = document.querySelector('.scale__control--bigger');
-const DOWN_SCALE = document.querySelector('.scale__control--smaller');
-const SCALE_VALUE = document.querySelector('.scale__control--value');
-
-const IMG_UPLOAD_PREVIEW = document.querySelector('.img-upload__preview');
-
-const IMG_UPLOAD_EFFECTS = document.querySelector('.img-upload__effects');
-const EFFECTS_LIST = IMG_UPLOAD_EFFECTS.querySelector('.effects__list');
-
-const EFFECT_NONE = EFFECTS_LIST.querySelector('.effects__item #effect-none');
-const EFFECT_CHROME = EFFECTS_LIST.querySelector('.effects__item #effect-chrome');
-const EFFECT_SEPIA = EFFECTS_LIST.querySelector('.effects__item #effect-sepia');
-const EFFECT_MARVIN = EFFECTS_LIST.querySelector('.effects__item #effect-marvin');
-const EFFECT_PHOBOS = EFFECTS_LIST.querySelector('.effects__item #effect-phobos');
-const EFFECT_HEAT = EFFECTS_LIST.querySelector('.effects__item #effect-heat');
-
-
-/* global noUiSlider:readonly */
-noUiSlider.create(SLIDER_ELEMENT, {
-  range: {
-    min: 0,
-    max: 1,
-  },
-  start: 1,
-  step: 0.1,
-  connect: 'lower',
-  format: {
-    to: function (value) {
-      if (Number.isInteger(value)) {
-        return value.toFixed(0);
-      }
-      return value.toFixed(1);
-    },
-    from: function (value) {
-      return parseFloat(value);
-    },
-  },
-})
-
-EFFECT_CHROME.addEventListener('change', (evt) => {
-  if (evt.target.checked) {
-    SLIDER_ELEMENT.noUiSlider.updateOptions({
-      range: {
-        min: 0,
-        max: 1,
-      },
-      start: 1,
-      step: 0.1,
-    });
-  }
-});
-
-EFFECT_SEPIA.addEventListener('change', (evt) => {
-  if (evt.target.checked) {
-    SLIDER_ELEMENT.noUiSlider.updateOptions({
-      range: {
-        min: 0,
-        max: 1,
-      },
-      start: 1,
-      step: 0.1,
-    });
-  }
-});
-
-EFFECT_MARVIN.addEventListener('change', (evt) => {
-  if (evt.target.checked) {
-    SLIDER_ELEMENT.noUiSlider.updateOptions({
-      range: {
-        min: 0,
-        max: 100,
-      },
-      start: 100,
-      step: 1,
-    });
-  }
-});
-
-EFFECT_PHOBOS.addEventListener('change', (evt) => {
-  if (evt.target.checked) {
-    SLIDER_ELEMENT.noUiSlider.updateOptions({
-      range: {
-        min: 0,
-        max: 3,
-      },
-      start: 3,
-      step: 0.1,
-    });
-  }
-});
-
-EFFECT_HEAT.addEventListener('change', (evt) => {
-  if (evt.target.checked) {
-    SLIDER_ELEMENT.noUiSlider.updateOptions({
-      range: {
-        min: 0,
-        max: 3,
-      },
-      start: 3,
-      step: 0.1,
-    });
-  }
-});
-//===============================================================================
-EFFECT_NONE.addEventListener('click', function () {
-
-  SLIDER_ELEMENT.noUiSlider.on('update', () => {
-    IMG_UPLOAD_PREVIEW.style.filter = 'none';
-
-  });
-
-  IMG_UPLOAD_PREVIEW.classList.remove('effects__preview--heat', 'effects__preview--chrome', 'effects__preview--sepia',
-    'effects__preview--marvin', 'effects__preview--phobos');
-
-  IMG_UPLOAD_PREVIEW.classList.add('effects__preview--none');
-});
-
-
-EFFECT_SEPIA.addEventListener('click', function () {
-  IMG_UPLOAD_PREVIEW.style.filter = 'sepia(1)';
-
-  SLIDER_ELEMENT.noUiSlider.on('update', (values, handle) => {
-    IMG_UPLOAD_PREVIEW.style.filter = `sepia(${values[handle]})`;
-  });
-
-  IMG_UPLOAD_PREVIEW.classList.remove('effects__preview--heat', 'effects__preview--chrome', 'effects__preview--none',
-    'effects__preview--marvin', 'effects__preview--phobos');
-
-  IMG_UPLOAD_PREVIEW.classList.add('effects__preview--sepia');
-});
-
-EFFECT_CHROME.addEventListener('click', function () {
-  IMG_UPLOAD_PREVIEW.style.filter = 'grayscale(1)';
-
-  SLIDER_ELEMENT.noUiSlider.on('update', (values, handle) => {
-    IMG_UPLOAD_PREVIEW.style.filter = `grayscale(${values[handle]})`;
-  });
-
-  IMG_UPLOAD_PREVIEW.classList.remove('effects__preview--heat', 'effects__preview--sepia', 'effects__preview--none',
-    'effects__preview--marvin', 'effects__preview--phobos');
-
-  IMG_UPLOAD_PREVIEW.classList.add('effects__preview--chrome');
-});
-
-EFFECT_MARVIN.addEventListener('click', function () {
-  IMG_UPLOAD_PREVIEW.style.filter = 'invert(100%)';
-
-  SLIDER_ELEMENT.noUiSlider.on('update', (values, handle) => {
-    IMG_UPLOAD_PREVIEW.style.filter = `invert(${values[handle]}%)`;
-  });
-
-  IMG_UPLOAD_PREVIEW.classList.remove('effects__preview--heat', 'effects__preview--sepia', 'effects__preview--none',
-    'effects__preview--chrome', 'effects__preview--phobos');
-
-  IMG_UPLOAD_PREVIEW.classList.add('effects__preview--marvin');
-});
-
-EFFECT_PHOBOS.addEventListener('click', function () {
-  IMG_UPLOAD_PREVIEW.style.filter = 'blur(3px)';
-
-  SLIDER_ELEMENT.noUiSlider.on('update', (values, handle) => {
-    IMG_UPLOAD_PREVIEW.style.filter = `blur(${values[handle]}px)`;
-  });
-
-  IMG_UPLOAD_PREVIEW.classList.remove('effects__preview--heat', 'effects__preview--sepia', 'effects__preview--none',
-    'effects__preview--chrome', 'effects__preview--marvin');
-
-  IMG_UPLOAD_PREVIEW.classList.add('effects__preview--phobos');
-});
-
-EFFECT_HEAT.addEventListener('click', function () {
-  IMG_UPLOAD_PREVIEW.style.filter = 'brightness(3)';
-
-  SLIDER_ELEMENT.noUiSlider.on('update', (values, handle) => {
-    IMG_UPLOAD_PREVIEW.style.filter = `brightness(${values[handle]})`;
-  });
-
-  IMG_UPLOAD_PREVIEW.classList.remove('effects__preview--phobos', 'effects__preview--sepia', 'effects__preview--none',
-    'effects__preview--chrome', 'effects__preview--marvin');
-
-  IMG_UPLOAD_PREVIEW.classList.add('effects__preview--heat');
-});
-
-//=======================================
-let counter = 100;
-SCALE_VALUE.value = counter + '%';
-
-UP_SCALE.addEventListener('click', function () {
-  if (counter <= 99) {
-
-    counter += 25;
-    SCALE_VALUE.value = counter + '%';
-
-    IMG_UPLOAD_PREVIEW.style.transform = `scale(0.${counter})`;
-    if (counter == 100) {
-      IMG_UPLOAD_PREVIEW.style.transform = 'scale(1)';
-    }
-  }
-
-});
-
-DOWN_SCALE.addEventListener('click', function () {
-  if (counter >= 26) {
-    counter -= 25;
-    SCALE_VALUE.value = counter + '%';
-    IMG_UPLOAD_PREVIEW.style.transform = `scale(0.${counter})`;
-  }
-
-});
-//===========================================
-
 export {
   getObjectDataPic
